% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{align_koma_attr}
\alias{align_koma_attr}
\title{Align a koma_ts attribute or metadata series to another time series}
\usage{
align_koma_attr(x, attr = NULL, template)
}
\arguments{
\item{x}{A \code{koma_ts} object or metadata series.}

\item{attr}{Name of the attribute to align. Only used when \code{x} is \code{koma_ts}.}

\item{template}{A time series whose \code{tsp} should be used.}
}
\value{
The aligned attribute value.
}
\description{
Align a koma_ts attribute or metadata series to another time series
}
