% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{as_list}
\alias{as_list}
\alias{as_list.mts}
\alias{as_list.list}
\title{Convert an mets koma_ts object to a list}
\usage{
as_list(x, ...)

\method{as_list}{mts}(x, ...)

\method{as_list}{list}(x, ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{Additional arguments.}
}
\value{
A list with ets koma_ts objects.
}
\description{
Convert an mets koma_ts object to a list
}
\examples{
x <- as_mets(list(
  y = as_ets(ts(1:8, start = c(2020, 1), frequency = 4)),
  z = as_ets(ts(11:18, start = c(2020, 1), frequency = 4))
))
as_list(x)
}
