% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{attach_color_code}
\alias{attach_color_code}
\title{Attach Color Codes to Data Frame Based on Status}
\usage{
attach_color_code(df_long, marker_color, status_column)
}
\arguments{
\item{df_long}{A data frame containing a column specified by \code{status_column}
which indicates the status of each sample.}

\item{marker_color}{A named list or other key-value mapping structure where
the names or keys correspond to statuses and the values correspond to color
codes.}

\item{status_column}{The name of the column in \code{df_long} that contains the
status information.}
}
\value{
A data frame identical to \code{df_long}, but with an additional column
\code{color_code} which contains the color codes mapped from the \code{status_column}
based on the \code{marker_color} mapping.
}
\description{
This function maps the \code{status_column} values in the provided data frame to
color codes, based on a specified color mapping, and attaches these color
codes as a new column \code{color_code} in the data frame.
}
\keyword{internal}
