% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{calculate_eq_weights}
\alias{calculate_eq_weights}
\title{Calculate Dynamic Weights for Identity Equations}
\usage{
calculate_eq_weights(ts_data, iden, start, end)
}
\arguments{
\item{ts_data}{A named list of \code{stats::ts} objects. It should contain the
time series data for both the aggregate value and its components. Ensure that
\code{ts_data} contains the time series listed in \code{iden}.}

\item{start}{A numeric value representing the start date for the time series
data subset. This value should be in the numeric format used by the function
\code{dates_to_num}.}

\item{end}{A numeric value representing the end date for the time series data
subset, in the same numeric format as \code{start}.}
}
\value{
A list of \code{stats::ts} objects containing the calculated weights for
the given aggregate value and its components. The top-level name of the
list is the aggregate value, and its sub-level names correspond to the
component names.
}
\description{
Computes the weights required for identity equations using time series data.
The weights represent how components collectively sum up to or are related to
an aggregate value over the specified time period.
}
\keyword{internal}
