% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{calculate_error}
\alias{calculate_error}
\title{Calculate the Forecast Error}
\usage{
calculate_error(forecasts, realized, variables)
}
\arguments{
\item{forecasts}{A mets-object containing the forecasts.}

\item{realized}{A mets-object containing the true values.}

\item{variables}{A character vector of name(s) of the stochastic
endogenous variable for which the forecast error(s) should be calculated.}
}
\value{
List of errors
}
\description{
This function calculates the the desired forecast errors for the
desired variables given the true values and the forecasts
}
\keyword{internal}
