% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_sem.R
\name{compute_forecast_values}
\alias{compute_forecast_values}
\title{Compute Forecast Based on Companion Matrix and Reduced Form}
\usage{
compute_forecast_values(
  companion_matrix,
  reduced_form,
  y_matrix,
  forecast_x_matrix,
  horizon,
  freq,
  start_forecast,
  endogenous_variables,
  restrictions
)
}
\arguments{
\item{companion_matrix}{A list containing components of the companion matrix.}

\item{reduced_form}{A list containing the reduced form components.}

\item{y_matrix}{Matrix of the dependent variable time series data.}

\item{forecast_x_matrix}{A matrix with forecasting data for exogenous
variables.}

\item{horizon}{Forecasting horizon, specifying the number of periods.}

\item{freq}{Frequency of the time series data.}

\item{start_forecast}{A vector of format \code{c(YEAR, QUARTER)} representing
the start date of the forecast.}

\item{endogenous_variables}{A character vector containing the names of
endogenous variables.}

\item{restrictions}{List of model constraints. Default is empty.}
}
\value{
A matrix with the base forecast of Y.
}
\description{
This function calculates the forecast for a given set of parameters and data.
It computes baseline forecasts and applies certain restrictions if specified.
}
\keyword{internal}
