% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{concat}
\alias{concat}
\title{Concatenate two time series}
\usage{
concat(x, y, ...)
}
\arguments{
\item{x}{A \code{koma_ts} object.}

\item{y}{A \code{koma_ts} object to be concatenated to x.}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
A \code{koma_ts} object containing \code{x} followed by \code{y}.
}
\description{
Concatenate two time series
}
\examples{
x <- as_ets(ts(c(100, 101, 102), start = c(2020, 1), frequency = 4))
y <- as_ets(ts(c(103, 104), start = c(2020, 4), frequency = 4))
concat(x, y)
}
