% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ragged_edge.R
\name{conditional_fill}
\alias{conditional_fill}
\title{Conditional Fill and Forecast for Time Series Data}
\usage{
conditional_fill(ts_data, sys_eq, dates, estimates, point_forecast)
}
\arguments{
\item{ts_data}{Time-series data set for the estimation.}

\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{estimates}{Optional. A \code{koma_estimate} object
(see \code{\link{estimate}}) containing the estimates of the previously
estimated simultaneous equations model. Use this parameter when some
equations of the system need to be re-estimated.}

\item{point_forecast}{A list of options used when ragged edge is filled
that contains:
\itemize{
\item active: Determines the type of forecast generated.
If TRUE, a density forecast is created. If FALSE, a point forecast is
returned. Default is TRUE.
\item central_tendency A character string indicating which central tendency
measure ("mean" or "median") to use when point_forecast$active is TRUE.
Default is "mean".
}}
}
\value{
An extended time series list containing the filled and forecasted
time series up to the current quarter.
}
\description{
This function fills missing observations and extends the time series data
up to and including the current quarter. The function conditionally forecasts
based on the estimates and realized observations.
}
\keyword{internal}
