% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_matrix_construction.R
\name{construct_beta_matrix}
\alias{construct_beta_matrix}
\title{Construct Beta Matrix}
\usage{
construct_beta_matrix(equations, exogenous_variables)
}
\arguments{
\item{equations}{A character string containing the system of equations,
where equations are separated by commas.}

\item{exogenous_variables}{A character vector representing the
exogenous variables.}
}
\value{
Beta matrix.
}
\description{
Constructs the B matrix based on the given system of equations and the
exogenous variables.
}
\keyword{internal}
