% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companion_matrix.R
\name{construct_companion_matrix}
\alias{construct_companion_matrix}
\title{Construct Companion Matrices for Dynamic SEM}
\usage{
construct_companion_matrix(posterior, exogenous_variables)
}
\arguments{
\item{posterior}{A list containing gamma_matrix, beta_matrix,  sigma_matrix
and phi_matrix derived from the LIBA estimates.}

\item{exogenous_variables}{A character vector of names of exogenous
variables.}
}
\value{
A list containing matrices in companion form: phi_matrix,
beta_tilde_matrix, gamma_matrix, c_matrix, and values n and p.
}
\description{
Construct Companion Matrices for Dynamic SEM
}
\keyword{internal}
