% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_matrix_construction.R
\name{construct_gamma_matrix}
\alias{construct_gamma_matrix}
\title{Construct Gamma Matrix}
\usage{
construct_gamma_matrix(equations, endogenous_variables)
}
\arguments{
\item{equations}{A character string containing the system of equations,
where equations are separated by commas.}

\item{endogenous_variables}{A character vector representing the
endogenous variables.}
}
\value{
A Gamma matrix.
}
\description{
Constructs a Gamma matrix based on the given system of equations and
endogenous variables.

This function constructs a Gamma matrix, which is a diagonal matrix
representing the coefficients of endogenous variables in the system
of equations. The Gamma matrix is used in various econometric models.
It also returns the constructed weights and parameters.
}
\keyword{internal}
