% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_posterior.R
\name{construct_posterior}
\alias{construct_posterior}
\title{Construct Posterior Matrices from LIBA Estimates}
\usage{
construct_posterior(sys_eq, estimate)
}
\arguments{
\item{sys_eq}{A list containing detailed components of the
system of equations such as \code{equations}, \code{endogenous_variables},
\code{identities}, \code{character_gamma_matrix}, and \code{character_beta_matrix}.}

\item{estimate}{A draw that contains beta, gamma and omega tilde estimates.}
}
\value{
A list containing gamma_matrix, beta_matrix, sigma_matrix and
phi_matrix derived from the LIBA estimates.
}
\description{
This function constructs the posterior matrices using the LIBA estimates.
It primarily focuses on the posterior median.
}
\keyword{internal}
