% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companion_matrix.R
\name{construct_reduced_form}
\alias{construct_reduced_form}
\title{Construct Reduced Form Parameters from Companion Matrix}
\usage{
construct_reduced_form(companion_matrix)
}
\arguments{
\item{companion_matrix}{A list containing the matrices beta_tilde_matrix,
phi_matrix, gamma_matrix, and c_matrix used to calculate the reduced form
parameters.}
}
\value{
A list containing the matrices companion_pi, companion_theta, and
companion_d that represent the reduced form parameters
}
\description{
This function computes the reduced form parameters companion_pi,
companion_theta, and companion_d using the provided companion_matrix.
}
\keyword{internal}
