% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_construction.R
\name{construct_y_matrix_j}
\alias{construct_y_matrix_j}
\title{Constructs a matrix of endogenous variables appearing in equation j}
\usage{
construct_y_matrix_j(y_matrix, character_gamma_matrix, jx)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}
}
\value{
If equation \eqn{j} contains \eqn{\gamma} parameters, the function
returns a subset of \eqn{Y}.
If there are no gamma parameters, the function returns NA.
}
\description{
This extracts a \eqn{(T x n_j)} matrix of endogenous variables appearing in
equation \eqn{j}, with \eqn{T} being the number of observations and
\eqn{n_j} the number of endogenous variables in equation \eqn{j}.
}
\keyword{internal}
