% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_construction.R
\name{construct_z_matrix_j}
\alias{construct_z_matrix_j}
\title{Construct Z_j}
\usage{
construct_z_matrix_j(gamma_parameters_j, y_matrix, y_matrix_j, jx)
}
\arguments{
\item{gamma_parameters_j}{A \eqn{(n_j \times 1)} matrix with the parameters
of the \eqn{\gamma} matrix, where \eqn{n_j} is the number of endogenous
variables in equation \eqn{j}.}

\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{y_matrix_j}{A \eqn{(T \times n_j)} matrix of endogenous variables
appearing in equation \eqn{j}, with \eqn{T} being the number of observations
and \eqn{n_j} the number of endogenous variables in equation \eqn{j}.}

\item{jx}{The index of equation \eqn{j}.}
}
\value{
\eqn{Z_j} matrix with dimensions \eqn{T \times 2}
}
\description{
\eqn{Z_j = [ \, y_j - Y_j * \gamma_j, Y_j ] \,}
}
\keyword{internal}
