% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lagged_variables.R
\name{create_lagged_variables}
\alias{create_lagged_variables}
\title{Create Lagged Variables for Time Series Data}
\usage{
create_lagged_variables(
  ts_data,
  endogenous_variables,
  exogenous_variables,
  predetermined_variables
)
}
\arguments{
\item{ts_data}{A named list of time series objects.}

\item{endogenous_variables}{A character vector of names of endogenous
variables.}

\item{exogenous_variables}{A character vector of names of exogenous
variables.}

\item{predetermined_variables}{A character vector of names of predetermined
variables, which contains lag information in the format "var.L(k)" where
"var" is the variable name and "k" is the lag (e.g., "var.L(1)" for the
first lag of the variable "var").}
}
\value{
A named list of time series objects with the lagged variables
added according to the predetermined pattern.
}
\description{
This function creates lagged variables for the given time series data,
considering the provided endogenous and exogenous variables along with
predetermined variables. The lagged variables are created based on the
pattern of the predetermined variables.
}
\keyword{internal}
