% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_processing.R
\name{detect_lag}
\alias{detect_lag}
\title{Detect Lag Number from Variable Name}
\usage{
detect_lag(variable_name)
}
\arguments{
\item{variable_name}{A character string representing the variable name.}
}
\value{
An integer representing the detected lag number from the variable
name. Returns NA if no lag detected.
}
\description{
This function takes a variable name as input and extracts the lag number
from it.
The lag number is assumed to be a positive integer preceded by the letter
'L'.
}
\keyword{internal}
