% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm_informative.R
\name{draw_gamma_j_informative}
\alias{draw_gamma_j_informative}
\title{Draw gamma parameters for equation j}
\usage{
draw_gamma_j_informative(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx,
  gamma_jw,
  tau,
  cholesky_of_inverse_hessian,
  omega_jw,
  theta_jw,
  priors_j
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}

\item{gamma_jw}{A \eqn{(n_j \times 1)} vector with the parameters
of the gamma matrix, where \eqn{n_j} is the number of endogenous variables in
equation \eqn{j}.}

\item{tau}{A tuning scalar \eqn{\tau} to adjust the acceptance rate.}

\item{cholesky_of_inverse_hessian}{The Cholesky factor \eqn{L} of the
inverse Hessian matrix \eqn{M^{-1}} used to generate candidate draws.}

\item{omega_jw}{\eqn{{\Omega}_j^{(w)}}}

\item{theta_jw}{A \eqn{(k \times nj+1)}, where \eqn{n_j} is the number of
endogenous variables in equation \eqn{j}.}
}
\value{
A \eqn{(n_j \times 1)} matrix with the either accepted candidate or
previous gamma parameters. Returns 0 if there are no endogenous
variables in equation \eqn{j}.
}
\description{
\code{draw_gamma_j_informative} draws the \eqn{\gamma} parameters from a given
posterior distribution for an equation \eqn{j}, in the Metropolis-Hastings
algorithm.
}
\keyword{internal}
