% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm.R
\name{draw_parameters_j}
\alias{draw_parameters_j}
\title{Draw Parameters for equation j}
\usage{
draw_parameters_j(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}
}
\value{
A list containing matrices for the saved draws of parameters and
additional diagnostic information.
}
\description{
\code{draw_parameters_j} simulates from the posterior of the model's parameters
for each equation \eqn{j} separately.
The posterior is simulated using a Metropolis-within-Gibbs sampling
procedure.
}
\details{
The sampler works as follows:
\enumerate{
\item Initialize sampler
\item Conditional on \eqn{\Omega^{(w-1)}} and the data draw
\eqn{delta_{\gamma}}.
\item Conditional on \eqn{delta_{\gamma}^{(w)}} draw \eqn{\Omega^{(w)}}.
\item Conditional on \eqn{Omega^{(w)}} and \eqn{\delta_{\gamma}^{(w)}} draw
\eqn{\delta_\beta^{(w)}}.
\item Go back to step 2.
}
}
\keyword{internal}
