% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sem.R
\name{estimate_sem}
\alias{estimate_sem}
\title{Estimate Parameters in a System of Equations}
\usage{
estimate_sem(sys_eq, y_matrix, x_matrix, eq_jx = NULL)
}
\arguments{
\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{eq_jx}{A numeric vector indicating the indices of the endogenous
equations \eqn{j} to be estimated. If NULL, all endogenous equations are
estimated. The vector should contain positive integers corresponding to the
positions of the equations within the \code{endogenous_variables} list.}
}
\value{
List of estimates for the endogenous variables.
}
\description{
This function estimates parameters in a given system of equations using
either a single thread or parallel computing.
}
\details{
This function provides the option for parallel computing through
the \code{future::plan()} function. For more details, see the future package
documentation:
https://cran.r-project.org/web/packages/future/future.pdf
}
\keyword{internal}
