% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates_from_draws.R
\name{extract_estimates_from_draws}
\alias{extract_estimates_from_draws}
\title{Extract Median Estimates}
\usage{
extract_estimates_from_draws(
  sys_eq,
  estimates,
  jx = NULL,
  central_tendency = "mean"
)
}
\arguments{
\item{sys_eq}{A list containing detailed components of the
system of equations such as \code{endogenous_variables},
\code{total_exogenous_variables}, \code{identity_equations}, \code{character_beta_matrix},
and \code{character_gamma_matrix}.}

\item{estimates}{A list of estimates, obtained from previous estimation
procedures.}

\item{jx}{If an index is set then the posterior is constructed for estimate
of draw jx, otherwise the posterior is constructed using the median of the
estimates.
The default is that the posterior is constructed using the median estimates.}

\item{central_tendency}{A character string indicating which central tendency
measure ("mean" or "median") to use.
Default is "mean". Active only when jx is NULL.}
}
\value{
A list containing:
\itemize{
\item \code{gamma_matrix}: Either central tendency estimates or specific draw
estimates for gamma values, depending on \code{jx}.
\item \code{beta_matrix}: Either central tendency estimates or specific draw
estimates for beta values, depending on \code{jx}.
\item \code{sigma_matrix}: Either central tendency estimates or specific draw
estimates for sigma values, depending on \code{jx}.
}
}
\description{
This function computes either central tendency estimates (median or mean) or
returns estimates from a specific draw for gamma, beta, and sigma matrices.
It is mainly used internally by the \code{construct_posterior} function.
}
\keyword{internal}
