% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ragged_edge.R
\name{fill_ragged_edge}
\alias{fill_ragged_edge}
\title{Fill Ragged Edges in Time Series Data}
\usage{
fill_ragged_edge(ts_data, sys_eq, exogenous_variables, dates, fill_method)
}
\arguments{
\item{ts_data}{Time-series data set for the estimation.}

\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{exogenous_variables}{A character vector of exogenous variables.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{fill_method}{Character string indicating which central tendency measure
("mean" or "median") to use when filling ragged edges.}
}
\value{
A list containing the updated time series data.
}
\description{
This function fills in the ragged edges in a time series data set using a
system of equations model. It iteratively detects edges, estimates the
model, and fills the unobserved series using a one-step ahead conditional
forecast until the time series is balanced.
}
\keyword{internal}
