% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{filter_by_attribute}
\alias{filter_by_attribute}
\title{Filter a koma_ts object by attribute value}
\usage{
filter_by_attribute(x, attribute, value, var = NULL, ...)
}
\arguments{
\item{x}{A koma_ts object or a list of koma_ts objects.}

\item{attribute}{The attribute name to filter by.}

\item{value}{The attribute value to match.}

\item{var}{An optional variable to filter by.}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
A koma_ts object or list with the matching series.
}
\description{
Filter a koma_ts object by attribute value
}
\examples{
x <- as_ets(
  ts(1:8, start = c(2020, 1), frequency = 4),
  series_type = "level",
  method = "diff_log"
)
filter_by_attribute(
  list(x = x),
  attribute = "series_type",
  value = "level"
)
}
