% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ticks.R
\name{find_extremas}
\alias{find_extremas}
\title{Find Extrema Values for Growth Rates and Levels}
\usage{
find_extremas(df_long)
}
\arguments{
\item{df_long}{A data frame in long format that includes 'data_type' as a
column, which should contain the types 'growth' and 'level' for
categorizing the data.}
}
\value{
A list with two elements 'growth' and 'level', each containing
a 'min' and 'max' sub-element. These represent the adjusted minimum
and maximum values for the growth rates and levels, respectively.
}
\description{
Computes the minimum and maximum values for growth rates and levels from
a given data frame, adjusting these values by a certain percentage to provide
a margin. This function is useful for determining the range of y-axis values
in plotting functions.
}
\keyword{internal}
