% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_sem.R
\name{forecast_sem}
\alias{forecast_sem}
\title{Generate Forecasts for a System of Equations}
\usage{
forecast_sem(
  sys_eq,
  estimates,
  restrictions,
  y_matrix,
  forecast_x_matrix,
  horizon,
  freq,
  forecast_dates,
  point_forecast
)
}
\arguments{
\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{estimates}{List of parameter estimates for the system.}

\item{restrictions}{List of model constraints. Default is empty.}

\item{y_matrix}{Matrix of the dependent variable time series data.}

\item{forecast_x_matrix}{A matrix with forecasting data for exogenous
variables.}

\item{horizon}{Forecasting horizon, specifying the number of periods.}

\item{freq}{Frequency of the time series data.}

\item{forecast_dates}{List containing 'start' and 'end' dates for forecast.}

\item{point_forecast}{A list that contains the following elements:
\itemize{
\item \code{active}: Determines the type of forecast generated.
If TRUE, a point forecast is created. If FALSE, a density forecast is
returned. Default is TRUE.
\item \code{central_tendency}: A character string indicating which
central tendency measure ("mean" or "median") to use for summary statistics.
Default is "mean".
}}
}
\value{
A list containing the forecast values.
}
\description{
This function performs forecasting based on the provided system of equations,
estimates, and other parameters. It supports both density and point
forecasting.
}
\keyword{internal}
