% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{format.koma_seq}
\alias{format.koma_seq}
\title{Format System of Equations}
\usage{
\method{format}{koma_seq}(x, ...)
}
\arguments{
\item{x}{An object of class \code{koma_seq}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A character vector of the formatted equations.
}
\description{
This function formats an object of class \code{koma_seq} for better readability.
It formats the equations to ensure proper spacing around operators and aligns
the equations for a cleaner display.
}
