% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_generation.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate Data for the Simultaneous Equations Model}
\usage{
generate_sample_data(
  sample_size,
  sample_start,
  burnin,
  gamma_matrix,
  beta_matrix,
  sigma_matrix,
  endogenous_variables,
  exogenous_variables,
  predetermined_variables
)
}
\arguments{
\item{sample_size}{The number of observations.}

\item{sample_start}{A vector of format \code{c(YEAR, QUARTER)} representing}

\item{burnin}{The number of observations to discard to mitigate the
dependency on starting values.}

\item{gamma_matrix}{A matrix that defines the relationships between the
variables.}

\item{beta_matrix}{A matrix that defines the coefficients of the equation.}

\item{sigma_matrix}{A matrix that defines the standard deviation of the
error terms.}

\item{endogenous_variables}{A vector of endogenous variables.}

\item{exogenous_variables}{A vector of exogenous variables.}

\item{predetermined_variables}{A vector of predetermined variables.}
}
\value{
A list containing three elements:
\itemize{
\item y_matrix: The dependent variable matrix with newly generated data.
\item x_matrix: The matrix of exogenous and predetermined variables with
newly generated data.
\item ts_data: A list containing time series in growth rates.
}
}
\description{
Compute Y and X matrix (data)
Use reduced form of the model to compute Y
\eqn{Y = X B \Gamma^{-1} + U \Gamma^{-1}}
}
\details{
It creates a sample of data where the relationship between the variables
is determined by the defined parameters.
}
