% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_tendency.R
\name{get_central_tendency}
\alias{get_central_tendency}
\title{Get Central Tendency Value from a Named List}
\usage{
get_central_tendency(central_tendency, quantiles)
}
\arguments{
\item{central_tendency}{A string, either "mean" or "median".}

\item{quantiles}{A named list containing central tendency values.}
}
\value{
The central tendency value from the list \code{quantiles} based on the
\code{central_tendency} argument.
}
\description{
This function takes a named list \code{quantiles} and a string \code{central_tendency}
indicating which central tendency measure (mean or median) to retrieve.
}
\keyword{internal}
