% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_settings.R
\name{get_default_gibbs_spec}
\alias{get_default_gibbs_spec}
\title{Default Gibbs Specfications}
\usage{
get_default_gibbs_spec()
}
\description{
Default Gibbs Specfications
}
\section{Gibbs Sampler Specifications}{

\itemize{
\item \code{ndraws}: Integer specifying the number of Gibbs sampler draws. Default is
2000.
\item \code{burnin_ratio}: Numeric specifying the ratio for the burn-in period.
Default is 0.5.
\item \code{nstore}: Integer specifying the frequency of stored draws. Default is 1.
\item \code{tau}: Numeric tuning parameter for enforcing an acceptance rate.
Default is 1.1.
}
}

\keyword{internal}
