% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{get_endogenous_variables}
\alias{get_endogenous_variables}
\title{Get endogenous variables from a system of equations}
\usage{
get_endogenous_variables(equations)
}
\arguments{
\item{equations}{A character string containing the system of equations,
where equations are separated by commas.}
}
\value{
A character vector of endogenous variables.
}
\description{
This function takes a system of equations and returns it's
endogenous variables.
}
\keyword{internal}
