% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{get_identities}
\alias{get_identities}
\title{Extract identities from equations}
\usage{
get_identities(equations, character_gamma_matrix, character_beta_matrix)
}
\arguments{
\item{equations}{A character string containing the system of equations,
where equations are separated by commas.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations, with potential entries
like "theta", to be used in the weights extraction process.
The dimensions of the matrix are \eqn{(T \times n)}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations, with potential entries
like "theta", to be used in the weights extraction process.
The dimensions of the matrix are \eqn{(k \times n)}, where \eqn{k} is the
number of exogenous variables and \eqn{n} the number of equations.}
}
\value{
A list of identities extracted from the input equations.
Each entry contains:
\itemize{
\item equation: The raw equation
\item components: The components extracted from the equation
\item weights: The corresponding weights for each component
}
}
\description{
This function filters out equations containing 'epsilon', then extracts
weights and identities based on \code{character_gamma_matrix} and
\code{character_beta_matrix}.
}
\keyword{internal}
