% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_generation.R
\name{get_lagged_values}
\alias{get_lagged_values}
\title{Extract Lagged Values from Y Matrix}
\usage{
get_lagged_values(predetermined_matrix, y_matrix, tx)
}
\arguments{
\item{predetermined_matrix}{The predetermined matrix to be populated with
lagged values.}

\item{y_matrix}{The y matrix containing the original values.}

\item{tx}{The time index for which the lagged values should be obtained.}
}
\value{
The updated predetermined matrix with lagged values.
}
\description{
This function populates the matrix containing the predetermined
variables, i.e. the lagged variables, with the lagged values
contained in the Y matrix.
}
\keyword{internal}
