% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_processing.R
\name{get_max_lag}
\alias{get_max_lag}
\title{Determine the maximum lag of variables}
\usage{
get_max_lag(variable_names)
}
\arguments{
\item{variable_names}{A character vector specifying the names of
the variables to analyze.}
}
\value{
The maximum lag among the variables. If there are no lags detected
NA is returned.
}
\description{
This function calculates the maximum lag among a set of variables.
}
\keyword{internal}
