% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ticks.R
\name{get_optimal_ticks}
\alias{get_optimal_ticks}
\title{Find Optimal Ticks for y Axis}
\usage{
get_optimal_ticks(df_long, x_range, whisker_data = NULL)
}
\arguments{
\item{df_long}{A data frame in long format that includes 'data_type' as a
column, which should contain the types 'growth' and 'level' for
categorizing the data.}

\item{x_range}{A list with x-axis range with start and end values of format
c(YEAR, QUARTER).}
}
\value{
A list with y1 (growth) values and text, y2 (level)
text and extremas.
}
\description{
Find Optimal Ticks for y Axis
}
\keyword{internal}
