% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_seq_weights}
\alias{get_seq_weights}
\title{Get System of Equations Weights}
\usage{
get_seq_weights(ts_data, identities, dates)
}
\arguments{
\item{ts_data}{A named list of \code{stats::ts} objects. This list should contain
the time series data for both the aggregate values and their components.}

\item{identities}{A named list of identities, each containing components
and their respective weights. The list should include elements with
"lhs_weight" in their names.}

\item{dates}{A list with start and end dates for dynamic weight calculations.
The dates list should have a sub-list \code{dynamic_weights} with \code{start} and
\code{end} entries, each being a numeric representation of a date, it should be
in the format produced by \code{dates_to_num}.}
}
\value{
The function updates the \code{identities} list with the calculated
dynamic weights and returns nothing explicitly.
}
\description{
Calculates dynamic weights for a system of equations and updates the
identities with these weights. The function processes a list of time
series data (\code{ts_data}), a list of identities (\code{identities}), and a set of
dates (\code{dates}) to compute and update weights dynamically.
}
\keyword{internal}
