% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{get_variables}
\alias{get_variables}
\title{Extract Variables from a System of Equations}
\usage{
get_variables(equations)
}
\arguments{
\item{equations}{A character vector containing the equations to extract
variables from.}
}
\value{
A list of character vectors, where each vector contains the variables
from a single equation.
}
\description{
This function identifies and extracts variable names from a set of
equations. It removes whitespace and splits the equations by common
operators such as \code{~}, \code{==}, \code{+}, \code{*}, and \code{-}. It filters out components
that represent numeric values, only returning variable names.
}
