% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{get_weight_variables}
\alias{get_weight_variables}
\title{Extract Weight Names from Equations}
\usage{
get_weight_variables(equations)
}
\arguments{
\item{equations}{A character string or vector containing the system of
equations. If a single string, equations should be separated by commas.}
}
\value{
A character vector of unique weight names that appeared within
parentheses in the input equations.
}
\description{
This function takes a vector of equations and extracts the names of the
variables that appear within parentheses.
}
\keyword{internal}
