% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{get_weights_and_comps}
\alias{get_weights_and_comps}
\title{Extract Component Weights from an Equation}
\usage{
get_weights_and_comps(raw_eq, weights_list)
}
\arguments{
\item{raw_eq}{A character string of a mathematical equation. The equation is
expected to have a format like "lhs == rhs", where the right-hand side
(rhs) might contain components combined using '+' or '-'.}

\item{weights_list}{A pre-constructed list containing character
representations of weights derived from both \code{character_gamma_matrix} and
\code{character_beta_matrix}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{equation}: The original equation provided as input.
\item \code{components}: A list of components extracted from the rhs of the
equation.
\item \code{weights}: A list of weights, corresponding to each component. If a
component doesn't have an explicit weight in the equation, its weight is
NULL.
}
}
\description{
Given a mathematical equation, this function identifies its individual
components and their associated weights. For components paired with a numeric
value via a \code{*}, that numeric value is considered its weight. Otherwise, the
weight is set to NULL.
}
\seealso{
\code{\link{get_identities}} for a function that uses this function's
outputs.
}
\keyword{internal}
