% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ticks.R
\name{get_x_ticks}
\alias{get_x_ticks}
\title{Generate Tick Text and Values for a Plot}
\usage{
get_x_ticks(df_long, fig = NULL)
}
\arguments{
\item{df_long}{A data frame in a long format, containing at least the columns
\code{data_type}, \code{color_code}, \code{value}, and \code{dates}. This data frame should
have a row for each date and data type, along with the corresponding value
and color code.}

\item{fig}{An optional existing plotly figure object to which new ticks will
be added. If provided, the function will integrate new tick text and values
with the existing ticks on the figure. Default is NULL, indicating no
existing figure.}
}
\value{
A list containing two elements: \code{ticktext} and \code{tickvals}. The
\code{ticktext} element is a character vector of HTML formatted strings for
each tick, showcasing the year and annual growth rate. The \code{tickvals} element
is an integer vector of years corresponding to each tick.
}
\description{
This function generates tick text and values based on a provided data frame
and an optional existing plot figure. The tick text and values are
specifically structured for showcasing annual growth rates on a timeline.
}
\keyword{internal}
