% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ticks.R
\name{get_y_ticks}
\alias{get_y_ticks}
\title{Calculate Y-Axis Tick Marks for Multiple Axes in Plotly}
\usage{
get_y_ticks(extremas, num_ticks, center_around)
}
\arguments{
\item{extremas}{A list with 'min' and 'max' values for each y-axis.}

\item{num_ticks}{The desired number of ticks on each y-axis.}

\item{center_around}{A numeric value to center the axis ranges.}
}
\value{
A list with dtick, min, and max for each y-axis, adjusted for
alignment.
}
\description{
This function computes the tick marks for y-axes on a Plotly graph, ensuring
that gridlines align across multiple y-axes by matching dtick ratios. It
adjusts the maximum range of the axes based on these ratios to align
gridlines correctly.
}
\details{
The function calculates tick mark spacing (\code{dtick}) and ratios, then adjusts
axis ranges to match the largest dtick ratio, ensuring gridlines align across
axes. It avoids reducing the axis range maximum to prevent cutting off data
points.
}
\references{
Adapted from the methodology outlined by Victor Bezak in the GitHub
repository:
https://github.com/VictorBezak/Plotly_Multi-Axes_Gridlines
}
\keyword{internal}
