% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{is_system_of_equations}
\alias{is_system_of_equations}
\title{Check if Object is a System of Equations}
\usage{
is_system_of_equations(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Logical. Returns \code{TRUE} if the object inherits from the class
\code{koma_seq}, and \code{FALSE} otherwise.
}
\description{
This function checks if the given object inherits from the class \code{koma_seq},
indicating that it represents a system of equations.
}
