% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{iterate_n_periods}
\alias{iterate_n_periods}
\title{Add Periods to Numeric Date Representation}
\usage{
iterate_n_periods(x, n, frequency)
}
\arguments{
\item{x}{A numeric value or a 2-element vector (year, period). The vector
should represent a year and a period (e.g., quarter or month).}

\item{n}{The number of periods to add. This can be positive or negative.}

\item{frequency}{A numeric value indicating the number of periods in a year
(e.g., 1, for years, 4 for quarters, 12 for months).}
}
\value{
A numeric representation of the date after adding the periods.
}
\description{
This function iterates a numeric date representation by a given number of
periods (e.g., years, quarters or months). It accepts a date in either vector
format (year, period) or its numeric representation and adds the specified
number of periods to it.
}
\keyword{internal}
