% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{koma_ts}
\alias{koma_ts}
\alias{ets}
\alias{as_ets}
\alias{is_ets}
\title{Extended Time-Series Object}
\usage{
ets(
  data = NA,
  start = NULL,
  end = NULL,
  frequency = NULL,
  deltat = NULL,
  ts.eps = getOption("ts.eps"),
  ...
)

as_ets(x = stats::ts(), ...)

is_ets(x)
}
\arguments{
\item{data}{a vector or matrix of the observed time-series
    values. A data frame will be coerced to a numeric matrix via
    \code{data.matrix}.  (See also \sQuote{Details}.)}

\item{start}{the time of the first observation.  Either a single
    number or a vector of two numbers (the second of which is an integer),
    which specify a natural time
    unit and a (1-based) number of samples into the time unit.  See
    the examples for the use of the second form.}

\item{end}{the time of the last observation, specified in the same way
    as \code{start}.}

\item{frequency}{the number of observations per unit of time.}

\item{deltat}{the fraction of the sampling period between successive
    observations; e.g., 1/12 for monthly data.  Only one of
    \code{frequency} or \code{deltat} should be provided.}

\item{ts.eps}{time series comparison tolerance.  Frequencies are
    considered equal if their absolute difference is less than
    \code{ts.eps}.}

\item{...}{Additional attributes.}

\item{x}{An object to be coerced / checked.}
}
\value{
A koma_ts object.

A koma_ts object.

TRUE if the object is of class \code{koma_ts}, otherwise FALSE.
}
\description{
The function \code{\link{ets}} is used to create an extended time-series
(ets) object. Any additional attribute passed is saved as such in the ets.

\code{\link{as_ets}} and \code{\link{is_ets}} coerce an object to a time-
series and test whether an object is a time series of class \code{koma_ts}.
}
\seealso{
\code{\link[stats]{ts}}
}
