% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{level}
\alias{level}
\title{Compute the level for a time series}
\usage{
level(x, ...)
}
\arguments{
\item{x}{A time series object. Supported classes are \code{ts} and
\code{ets} (a subclass of \code{ts}).}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
An \code{ets} object.
}
\description{
Required attributes of the input time series are:
\itemize{
\item \code{series_type} - "rate" or "level"
\item \code{method} - "percentage", "diff_log", "none", or an expression
}
}
\examples{
x <- ets(c(0.3, 0.1, 0.2, -0.1), series_type = "rate", method = "percentage")
level(x)

}
