% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{merge_koma_attrs}
\alias{merge_koma_attrs}
\title{Merge custom metadata from two koma_ts objects}
\usage{
merge_koma_attrs(e1, e2 = NULL, op = NULL, template = NULL)
}
\arguments{
\item{e1}{First operand.}

\item{e2}{Second operand.}

\item{op}{Name of the operation.}

\item{template}{Result template used for alignment.}
}
\value{
A named list of merged custom attributes.
}
\description{
Merge custom metadata from two koma_ts objects
}
