% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{multivariate_norm}
\alias{multivariate_norm}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
multivariate_norm(n = 1, mu, sigma, tol = 1e-06, empirical = FALSE)
}
\arguments{
\item{n}{The number of samples required.}

\item{mu}{A vector giving the means of the variables.}

\item{sigma}{A positive-definite symmetric matrix specifying the covariance
matrix of the variables.}

\item{tol}{Tolerance (relative to the largest variance) for numerical lack
of positive-definiteness in \code{sigma}.}

\item{empirical}{Logical. If true, mu and sigma specify the empirical not
population mean and covariance matrix.}
}
\value{
If \code{n = 1} a vector of the same length as \code{mu}, otherwise
an \code{n} by \code{length(mu)} matrix with one sample in each row.
}
\description{
Produces one or more samples from the specified multivariate normal
distribution.
}
\details{
The matrix decomposition is done vai \code{eigen}; although a Choleski
decomposition might be faster, the eigendecomposition is stabler.
}
\examples{
#'
\dontrun{
sigma <- matrix(c(10, 3, 3, 2), 2, 2)
sigma
var(multivariate_norm(n = 1000, rep(0, 2), sigma))
var(multivariate_norm(n = 1000, rep(0, 2), sigma, empirical = TRUE))
}
}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics
with S.  Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\seealso{
\code{\link[=rnorm]{rnorm()}}
}
\keyword{internal}
