% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{multivariate_norm_pdf}
\alias{multivariate_norm_pdf}
\title{Evaluate Multivariate Normal Density at x}
\usage{
multivariate_norm_pdf(x, mu, sigma)
}
\arguments{
\item{x}{A vector at which to evaluate the density.}

\item{mu}{A vector giving the means of the variables.}

\item{sigma}{A positive-definite symmetric matrix specifying the covariance
matrix of the variables.}
}
\value{
The density of the multivariate normal distribution at \code{x}.
}
\description{
Evaluate Multivariate Normal Density at x
}
\keyword{internal}
