% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{normalize_quantile_probs}
\alias{normalize_quantile_probs}
\title{Normalize Quantile Probabilities}
\usage{
normalize_quantile_probs(fan_quantiles)
}
\arguments{
\item{fan_quantiles}{Numeric probabilities in (0, 1] or percentages in
\code{[0, 100]}.}
}
\value{
A numeric vector of probabilities in (0, 1].
}
\description{
Normalize Quantile Probabilities
}
\keyword{internal}
