% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{num_to_dates}
\alias{num_to_dates}
\title{Convert Numeric to Date Format}
\usage{
num_to_dates(x, frequency, ...)
}
\arguments{
\item{x}{A numeric value or list of numeric values representing the date(s).
If \code{x} is a single numeric value, it should be in the format produced by
\code{dates_to_num}.}

\item{frequency}{A numeric value indicating the number of periods in a year
(e.g., 1 for years, 4 for quarters, 12 for months).}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A numeric vector of length 2 (year, period) or a list of such
vectors.
}
\description{
This function converts a numeric representation of a date back to a vector
or list of vectors with year and period (e.g., quarter or month).
The function supports input as a single numeric value, or a list of numeric
values. The numeric representation should match the format produced by
\code{dates_to_num}.
}
\keyword{internal}
