% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotli.R
\name{plotli}
\alias{plotli}
\title{Plotting Function}
\usage{
plotli(df_long, fig = NULL, theme = NULL, ...)
}
\arguments{
\item{df_long}{A data frame in long format, containing the data to be
plotted.}

\item{fig}{Optional. A Plotly figure object to update, otherwise creates a
new one.}

\item{theme}{List of default plot output parameters. Defaults to NULL, which
leads to \code{\link{init_koma_theme}} being called.
Please see the vignette for details about tweaking themes.}

\item{...}{Directly pass additional arguments to plotly::layout. For
documentation, see
\url{https://plotly.com/r/reference/#Layout_and_layout_style_objects}}
}
\value{
A Plotly figure object displaying the data with the specified formatting.
}
\description{
This function processes the data and creates a plot using Plotly and the
theme provided.
}
\references{
\url{https://plotly.com/r/reference/#Layout_and_layout_style_objects}
}
