% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transform.R
\name{prepare_data_to_plot}
\alias{prepare_data_to_plot}
\title{Prepare Data for Plotting}
\usage{
prepare_data_to_plot(mts_list, start)
}
\arguments{
\item{mts_list}{A list with named multivariate time series objects: growth,
level and / or growth_annual.}

\item{start}{A numeric value representing the forecast start date.}
}
\value{
A single tibble in long format combining growth rates, annualized
growth rates, and level data.
}
\description{
This function prepares time series data by converting the data from wide to
long format.
}
\keyword{internal}
