% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_settings.R
\name{print_gibbs_settings}
\alias{print_gibbs_settings}
\title{Display Current Gibbs Sampler Specifications}
\usage{
print_gibbs_settings(global_settings)
}
\value{
Prints a formatted message with the current Gibbs sampler
settings. This function does not return any values.
}
\description{
This function prints the current settings of the Gibbs sampler in a
readable format. It includes information such as the number of draws,
burn-in ratio, number of saved draws, and tuning parameter tau.
}
\keyword{internal}
