% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantile_names_from_probs}
\alias{quantile_names_from_probs}
\title{Build Quantile Names from Probabilities}
\usage{
quantile_names_from_probs(probs)
}
\arguments{
\item{probs}{Numeric probabilities.}
}
\value{
A character vector of quantile names.
}
\description{
Build Quantile Names from Probabilities
}
\keyword{internal}
