% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{rate}
\alias{rate}
\alias{rate.ts}
\alias{rate.list}
\alias{rate.mts}
\title{Compute the rate of change for a time series}
\usage{
rate(x, ...)

\method{rate}{ts}(x, ...)

\method{rate}{list}(x, ...)

\method{rate}{mts}(x, ...)
}
\arguments{
\item{x}{A time series object. Supported classes are \code{ts} and
\code{ets} (a subclass of \code{ts}).}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
An \code{ets} object.
}
\description{
Required attributes of the input time series are:
\itemize{
\item \code{series_type} - "rate" or "level"
\item \code{method} - "percentage", "diff_log", "none", or an expression
}
}
\examples{
x <- ets(1:10, series_type = "level", method = "diff_log")
rate(x)

}
