% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{rebase}
\alias{rebase}
\title{Rebase Time Series Data Relative to a Base Period}
\usage{
rebase(x, start, end, ...)
}
\arguments{
\item{x}{An ets object.}

\item{start}{the time of the first observation.  Either a single
    number or a vector of two numbers (the second of which is an integer),
    which specify a natural time
    unit and a (1-based) number of samples into the time unit.  See
    the examples for the use of the second form.}

\item{end}{the time of the last observation, specified in the same way
    as \code{start}.}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
An ets object with the level computed.
}
\description{
Calculates a rebased time series using a given date range as the base
period. The base period average is set to 100, and all other values are
scaled accordingly.
}
