% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{restore_koma_attrs}
\alias{restore_koma_attrs}
\title{Restore custom metadata onto a time series}
\usage{
restore_koma_attrs(x, attrs)
}
\arguments{
\item{x}{A time series object.}

\item{attrs}{A named list of custom attributes.}
}
\value{
A \code{koma_ts} object.
}
\description{
Restore custom metadata onto a time series
}
