% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_settings.R
\name{set_gibbs_settings}
\alias{set_gibbs_settings}
\title{Set Gibbs Settings}
\usage{
set_gibbs_settings(settings, equation_settings, envir = the)
}
\arguments{
\item{settings}{List with system-wide/global settings.}

\item{equation_settings}{List with equation specific settings.}

\item{envir}{Environment to create the gibbs_sampler in. Defaults to "the".}
}
\description{
Set Gibbs Settings
}
\keyword{internal}
