% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_settings.R
\name{set_gibbs_spec}
\alias{set_gibbs_spec}
\title{Set Gibbs Sampler Specifications}
\usage{
set_gibbs_spec(
  ndraws = NULL,
  burnin_ratio = NULL,
  nstore = NULL,
  tau = NULL,
  ...
)
}
\arguments{
\item{ndraws}{Integer specifying the number of Gibbs sampler draws.}

\item{burnin_ratio}{Numeric specifying the ratio for the burn-in period.}

\item{nstore}{Integer specifying the frequency of stored draws.}

\item{tau}{Numeric tuning parameter for enforcing an acceptance rate.}

\item{...}{Unused arguments that get discarded.}
}
\value{
Invisible NULL. The function updates \code{the$gibbs_sampler} in place.
}
\description{
This class updates the Gibbs sampler specifications in the global
environment variable \code{the$gibbs_sampler}.
}
\section{Gibbs Sampler Specifications}{

\itemize{
\item \code{ndraws}: Integer specifying the number of Gibbs sampler draws. Default is
2000.
\item \code{burnin_ratio}: Numeric specifying the ratio for the burn-in period.
Default is 0.5.
\item \code{nstore}: Integer specifying the frequency of stored draws. Default is 1.
\item \code{tau}: Numeric tuning parameter for enforcing an acceptance rate.
Default is 1.1.
}
}

\keyword{internal}
