% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{set_restrictions}
\alias{set_restrictions}
\title{Set Restrictions for Observed Variables in Time Series Data}
\usage{
set_restrictions(ts_data, variables_to_restrict, start, end)
}
\arguments{
\item{ts_data}{A list containing the time series data.}

\item{variables_to_restrict}{A vector of strings specifying the names of the
variables in \code{ts_data} that should be restricted.}

\item{start}{The starting date for the restrictions.}

\item{end}{The ending date for the restrictions.}
}
\value{
A named list of restrictions, where the names are the names of the
variables in \code{variables_to_restrict}. Each element is a list containing
\code{horizon}, a sequence from 1 to the length of \code{value_at_date}, and \code{value},
the truncated values of the variable between \code{start} and
\code{end}.
}
\description{
This function generates a list of restrictions for specified variables in a
time series data frame, based on their values between specified start and
end dates.
}
\keyword{internal}
