% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_open_economy}
\alias{small_open_economy}
\title{Quarterly time series of key macro variables for a small open economy (Switzerland)}
\format{
A \code{list} of 12 \code{ts} objects (quarterly):
\describe{
\item{consumption}{Private consumption \(C_t\).}
\item{investment}{Gross fixed capital formation \(I_t\).}
\item{exports}{Exports \(X_t\).}
\item{imports}{Imports \(M_t\).}
\item{prices}{Price level \(P_t\).}
\item{interest_rate}{Short‐term domestic interest rate \(R_t\).}
\item{gdp}{Equilibrium output \(GDP_t\).}
\item{domestic_demand}{Domestic demand \(D_t\).}
\item{world_gdp}{World GDP (exogenous) \(WGDP_t\).}
\item{interest_rate_germany}{German short‐term rate (exog) \eqn{R_{GE,t}}.}
\item{exchange_rate}{Nominal exchange rate (exog) \(ER_t\).}
\item{oil_price}{Oil price (exogenous) \(OP_t\).}
}
}
\source{
\url{https://www.seco.admin.ch/}, \url{https://kof.ethz.ch/}
}
\usage{
small_open_economy
}
\description{
Quarterly time series of key macro variables for a small open economy (Switzerland)
}
\keyword{datasets}
