% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{strip_koma_attrs}
\alias{strip_koma_attrs}
\title{Remove custom metadata from a koma_ts object}
\usage{
strip_koma_attrs(x)
}
\arguments{
\item{x}{A time series object.}
}
\value{
A plain \code{ts} object when \code{x} is \code{koma_ts}, otherwise \code{x}.
}
\description{
Remove custom metadata from a koma_ts object
}
