% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{summary.koma_forecast_hdr}
\alias{summary.koma_forecast_hdr}
\title{Summary for koma_forecast_hdr Objects}
\usage{
\method{summary}{koma_forecast_hdr}(object, ..., variables = NULL, horizon = NULL, digits = 3)
}
\arguments{
\item{object}{A \code{koma_forecast_hdr} object.}

\item{...}{Unused.}

\item{variables}{Optional character vector to filter variables.}

\item{horizon}{Optional numeric or character vector selecting forecast horizon.}

\item{digits}{Number of digits to round numeric values. Default is 3.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints HDR intervals for forecast horizons.
}
