% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{summary.koma_hdi}
\alias{summary.koma_hdi}
\title{Summary for koma_hdi Objects}
\usage{
\method{summary}{koma_hdi}(object, ..., digits = 3)
}
\arguments{
\item{object}{A \code{koma_hdi} object.}

\item{...}{Unused.}

\item{digits}{Number of digits to round numeric values. Default is 3.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints HDI intervals for a single numeric sample.
}
