% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{summary.koma_hdr}
\alias{summary.koma_hdr}
\title{Summary for koma_hdr Objects}
\usage{
\method{summary}{koma_hdr}(object, ..., digits = 3)
}
\arguments{
\item{object}{A \code{koma_hdr} object.}

\item{...}{Unused.}

\item{digits}{Number of digits to round numeric values. Default is 3.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints HDR intervals for a single numeric sample.
}
