% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{the}
\alias{the}
\title{Define the internal environment of the package}
\format{
An object of class \code{environment} of length 1.
}
\usage{
the
}
\description{
Define the internal environment of the package
}
\keyword{internal}
