% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transform.R
\name{to_long}
\alias{to_long}
\title{Convert Time Series from Wide to Long Format}
\usage{
to_long(mts, start)
}
\arguments{
\item{mts}{A multivariate time series object.}

\item{start}{A numeric value representing the forecast start date.}
}
\value{
A data frame in long format containing the original data along
with the sample status, dates, and frame identifiers.
}
\description{
This function returns the data in long format.
}
\keyword{internal}
