% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{type}
\alias{type}
\title{Get the type of a koma_ts object}
\usage{
type(x, type, var = NULL, ...)
}
\arguments{
\item{x}{A koma_ts object.}

\item{type}{The type to filter by.}

\item{var}{An optional variable to filter by.}

\item{...}{arguments passed to methods (unused for the default method).}
}
\value{
A koma_ts object with the specified type.
}
\description{
Get the type of a koma_ts object
}
