% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{update_anker}
\alias{update_anker}
\title{Update anker}
\usage{
update_anker(x, y)
}
\arguments{
\item{x}{Time series object of class \code{koma_ts} with realized / historical
data.}

\item{y}{Time series object of class \code{koma_ts} with forecast data.}
}
\value{
Return y of class \code{koma_ts} with updated anker.
}
\description{
This function updates the anker to the last observation of the realized
data.
}
\keyword{internal}
