% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_system_of_equations}
\alias{validate_system_of_equations}
\title{Validate Equations}
\usage{
validate_system_of_equations(x)
}
\value{
Returns object.
}
\description{
This function validates a character vector of equations, ensuring that they
adhere to a specific format.
The equations should be in the format "left_variable==right_variables", where
right variables can be separated by '+', '-', or '*'. The function checks for
valid equation structure, valid variable names, no duplicate regressors in a
single equation, and no duplicate dependent variables across all equations.
}
\keyword{internal}
