% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_identification.R
\name{vector_to_matrix}
\alias{vector_to_matrix}
\title{Convert Vector to Matrix}
\usage{
vector_to_matrix(
  transformation_matrix,
  parameters,
  constant_vector,
  nrow,
  ncol
)
}
\arguments{
\item{transformation_matrix}{A numeric matrix used for parameter
transformation.}

\item{parameters}{A numeric vector of parameters.}

\item{constant_vector}{A numeric vector for constant terms.}

\item{nrow}{Number of rows for the resulting matrix.}

\item{ncol}{Number of columns for the resulting matrix.}
}
\value{
A numeric matrix with specified row and column dimensions.
}
\description{
Given a transformation matrix, a vector of parameters, and a constant
vector, this function constructs a matrix with specified row and
column dimensions.
vector = transformation_matrix \%*\% parameters, constant_vector
where vector = vec(matrix)
}
\keyword{internal}
