% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_identification.R
\name{adjust_constant_vector}
\alias{adjust_constant_vector}
\title{Adjust Constant Vector Using Identity Weights}
\usage{
adjust_constant_vector(constant_vector, identity_weights)
}
\arguments{
\item{constant_vector}{A numeric vector representing the constant terms to
be adjusted.}

\item{identity_weights}{A list where each element corresponds to a set of
identity rules.}
}
\value{
The adjusted constant vector in numeric form with modified values
based on the identity weights.
}
\description{
Adjusts a given constant vector based on a list of identity weights.
The identity weights are used to replace specific elements in the
constant vector with corresponding numeric values.
}
