% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{as_list}
\alias{as_list}
\title{Convert an mets koma_ts object to a list}
\usage{
as_list(x, ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{Additional arguments.}
}
\value{
A list with ets koma_ts objects.
}
\description{
Convert an mets koma_ts object to a list
}
