% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{as_mets}
\alias{as_mets}
\title{Convert a list object with ets koma_ts objects to a koma_ts multivariate
time series (mets)}
\usage{
as_mets(x, ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{Additional arguments.}
}
\value{
A koma_ts multivariate time series object.
}
\description{
Convert a list object with ets koma_ts objects to a koma_ts multivariate
time series (mets)
}
