% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{concat}
\alias{concat}
\title{Concatenate two time series}
\usage{
concat(x, y, ...)
}
\arguments{
\item{x}{A \code{koma_ts} object.}

\item{y}{A \code{koma_ts} object to be concatenated to x.}

\item{...}{arguments passed to methods (unused for the default method).}
}
\description{
Concatenate two time series
}
