% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_posterior.R
\name{construct_phi}
\alias{construct_phi}
\title{Construct a Dynamic SEM (Structural Equation Model) Phi Matrix}
\usage{
construct_phi(sys_eq, beta_matrix)
}
\arguments{
\item{sys_eq}{A list containing the system of equations. Must
include \verb{$equations} with the equations of the system,
\verb{$endogenous_variables} with the names of the endogenous variables, and
\verb{$total_exogenous_variables} with the names of all exogenous variables.}

\item{beta_matrix}{A numeric matrix of beta coefficients corresponding to the
exogenous variables in the system of equations.}
}
\value{
A list of matrices representing the Phi(L) matrix for the given
system of equations.
}
\description{
This function constructs the Phi matrix from the given system of equations
and beta matrix. It processes the equations to find lagged endogenous
variables and uses them to build the Phi matrix, as used in dynamic SEM:
Y * Gamma = X_tilde B_tilde + Y(-1) * Phi(1) + ... + Y(-p) * Phi(L) + U.
}
\keyword{internal}
