% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_construction.R
\name{construct_theta_hat_j}
\alias{construct_theta_hat_j}
\title{Construct \eqn{\hat{\Theta}_j}}
\usage{
construct_theta_hat_j(x_matrix, z_matrix_j)
}
\arguments{
\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{z_matrix_j}{A \eqn{Z_j = y_j - Y_j * \gamma_j} matrix.}
}
\value{
\eqn{\hat{\Theta}_j} with dimensions \eqn{(k \times (1 + n_j))}.
}
\description{
Computes the OLS estimate for \eqn{\hat{\Theta}_j}, which
is a \eqn{(k \times (1 + n_j))} matrix, where \eqn{n_j} is the number
of endogenous variables in equation \eqn{j} and k is the number of
exogenous variables.
}
\details{
\eqn{\hat{\Theta}_j = (X'X)^{-1}X' Z_j}
}
\keyword{internal}
