% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm.R
\name{draw_omega_j}
\alias{draw_omega_j}
\title{Draw Omega from inverse Wishart distribution for equation j}
\usage{
draw_omega_j(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx,
  gamma_parameters_j
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}
}
\value{
List containing \eqn{{\tilde{\Omega}}_j^{(w)}} as \code{omega_tilde_jw}
and \eqn{\Omega_j^{(w)}} as \code{omega_jw}.
}
\description{
\code{draw_omega_j} draws the variance-covariance matrix \eqn{\tilde{\Omega}_j}
for each row of \eqn{[ u_j,  V_j ]}.
}
\keyword{internal}
