% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate the Simultaneous Equations Model (SEM)}
\usage{
estimate(
  ts_data,
  sys_eq,
  dates,
  ...,
  options = list(),
  point_forecast = NULL,
  estimates = NULL
)
}
\arguments{
\item{ts_data}{Time-series data set for the estimation.}

\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{...}{Additional parameters.}

\item{options}{Optional settings for modifying the Gibbs sampler
specifications for all equations.
See \link[=get_default_gibbs_spec]{Gibbs Sampler Specifications}.}

\item{point_forecast}{A list that contains the following elements:
\itemize{
\item \code{active}: Determines the type of forecast generated.
If TRUE, a point forecast is created. If FALSE, a density forecast is
returned. Default is TRUE.
\item \code{central_tendency}: A character string indicating which
central tendency measure ("mean" or "median") to use for summary statistics.
Default is "mean".
}}

\item{estimates}{Optional. A \code{koma_estimate} object
(see \code{\link{estimate}}) containing the estimates of the previously
estimated simultaneous equations model. Use this parameter when some
equations of the system need to be re-estimated.}
}
\value{
An object of class \code{koma_estimate}.

An object of class \code{koma_estimate}is a list containing the following
elements:
\describe{
\item{estimates}{The estimated parameters and other relevant information
obtained from the model.}
\item{sys_eq}{A \code{koma_seq} object containing details about the system of
equations used in the model.}
\item{ts_data}{The time-series data used for the estimation, with any \code{NA}
values removed and lagged variables created.}
\item{y_matrix}{The Y matrix constructed from the balanced data, used in
the estimation process.}
\item{x_matrix}{The X matrix constructed from the balanced data, used in
the estimation process.}
\item{gibbs_specifications}{The specifications used for the Gibbs
sampling.}
}
}
\description{
Estimate a system of simultaneous equations model (SEM) using a Bayesian
approach. This function incorporates Gibbs sampling and allows for
both density and point forecasts.
}
\section{Parallel}{

This function provides the option for parallel computing through
the \code{future::plan()} function.
For a detailed example on executing \code{estimate} in parallel, see the vignette:
\code{vignette("parallel")}.
For more details, see the \href{https://cran.r-project.org/web/packages/future/future.pdf}{future package documentation}.
}

\section{Gibbs Sampler Specifications}{

\itemize{
\item \code{ndraws}: Integer specifying the number of Gibbs sampler draws. Default is
2000.
\item \code{burnin_ratio}: Numeric specifying the ratio for the burn-in period.
Default is 0.5.
\item \code{nstore}: Integer specifying the frequency of stored draws. Default is 1.
\item \code{tau}: Numeric tuning parameter for enforcing an acceptance rate.
Default is 1.1.
}
}

\seealso{
\itemize{
\item To create a \code{koma_seq} object see \code{\link{system_of_equations}}.
\item For a comprehensive example of using \code{estimate}, see
\code{vignette("koma")}.
\item Related functions within the package that may be of interest:
\code{\link{forecast}}.
}
}
