% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast the Simultaneous Equations Model (SEM)}
\usage{
forecast(estimates, dates, ..., restrictions = NULL, point_forecast = NULL)
}
\arguments{
\item{estimates}{A \code{koma_estimate} object (\code{\link{estimate}}) containing
the estimates for the simultaneous equations model, as well as a list of time
series and a \code{koma_seq} object (\code{\link{system_of_equations}}) that were
used in the estimation.}

\item{dates}{Key-value list for date ranges in various model operations.}

\item{...}{Additional parameters.}

\item{restrictions}{List of model constraints. Default is empty.}

\item{point_forecast}{A list that contains the following elements:
\itemize{
\item \code{active}: Determines the type of forecast generated.
If TRUE, a point forecast is created. If FALSE, a density forecast is
returned. Default is TRUE.
\item \code{central_tendency}: A character string indicating which
central tendency measure ("mean" or "median") to use for summary statistics.
Default is "mean".
}}
}
\value{
An object of class \code{koma_forecast}.

An object of class \code{koma_forecast} is a list containing the following
elements:
\describe{
\item{mean}{Mean point forecasts as a list of time series of class
\code{koma_ts}.}
\item{median}{Median point forecasts as a list of time series of class
\code{koma_ts}.}
\item{quantiles}{A list of quantiles, where each element is named
according to the quantile (e.g., "5", "50", "95"), and contains the
forecasts for that quantile. This element is NULL if \code{quantiles = FALSE}.}
\item{ts_data}{Time-series data set used in forecasting.}
\item{y_matrix}{The Y matrix constructed from the balanced data up to the
current quarter, used for forecasting.}
\item{x_matrix}{The X matrix used for forecasting.}
}
}
\description{
This function produces forecasts for the SEM.
}
\details{
The \code{forecast} function for SEM uses the estimates from the \code{koma_estimate}
object to produce point forecasts or quantile forecasts based on the
\code{point_forecast} parameter. If \code{point_forecast$active} is \code{TRUE}, only point
forecasts are generated. If \code{FALSE}, quantile forecasts are generated and
included in the \code{quantiles} list.

Use the \code{\link{summary}} function to print a summary of the forecast
results.
The \code{\link{plot}} function can be used to visualize the forecasts and
prediction intervals.
}
\section{Parallel}{

This function provides the option for parallel computing through
the \code{future::plan()} function.
For a detailed example on executing \code{estimate} in parallel, see the vignette:
\code{vignette("parallel")}.
For more details, see the \href{https://cran.r-project.org/web/packages/future/future.pdf}{future package documentation}.
}

\seealso{
\itemize{
\item For a comprehensive example of using \code{forecast}, see
\code{vignette("koma")}.
\item Related functions within the package that may be of interest:
\code{\link{estimate}}.
}
}
