% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_sem.R
\name{forecast_draw}
\alias{forecast_draw}
\title{Generate a Forecast for a Single Draw}
\usage{
forecast_draw(
  sys_eq,
  estimates,
  jx,
  y_matrix,
  forecast_x_matrix,
  horizon,
  freq,
  forecast_dates,
  restrictions,
  state,
  central_tendency = NULL
)
}
\arguments{
\item{sys_eq}{A \code{koma_seq} object (\link{system_of_equations}) containing details
about the system of equations used in the model.}

\item{estimates}{List of parameter estimates for the system.}

\item{y_matrix}{Matrix of the dependent variable time series data.}

\item{forecast_x_matrix}{A matrix with forecasting data for exogenous
variables.}

\item{horizon}{Forecasting horizon, specifying the number of periods.}

\item{freq}{Frequency of the time series data.}

\item{forecast_dates}{List containing 'start' and 'end' dates for forecast.}

\item{restrictions}{List of model constraints. Default is empty.}

\item{state}{An environment used to share mutable state between function
calls, particularly for issuing warnings only once during the forecasting
process.}
}
\description{
This function computes a forecast for a single draw of the parameter
estimates, supporting both point forecasts (mean or median) and density
forecasts. It constructs the posterior distribution, companion matrix,
and reduced-form representation of the system before computing forecasts.
}
\keyword{internal}
