% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_within_gibbs_algorithm_informative.R
\name{initial_omega_j}
\alias{initial_omega_j}
\title{Compute initial omega for the jth equation}
\usage{
initial_omega_j(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx,
  gamma_jw
)
}
\arguments{
\item{y_matrix}{A \eqn{(T \times n)} matrix \eqn{Y}, where \eqn{T} is the
number of observations and \eqn{n} the number of equations, i.e. endogenous
variables.}

\item{x_matrix}{A \eqn{(T \times k)} matrix \eqn{X} of observations on
\eqn{k} exogenous variables.}

\item{character_gamma_matrix}{A matrix \eqn{\Gamma} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(T \times n)}, where \eqn{T} is the number of
observations and \eqn{n} the number of equations.}

\item{character_beta_matrix}{A matrix \eqn{\beta} that holds the
coefficients in character form for all equations. The dimensions of the
matrix are \eqn{(k \times n)}, where \eqn{k} is the number of
exogenous variables and \eqn{n} the number of equations.}

\item{jx}{The index of equation \eqn{j}.}

\item{gamma_jw}{A \eqn{(n_j \times 1)} vector with the parameters
of the gamma matrix, where \eqn{n_j} is the number of endogenous variables in
equation \eqn{j}.}
}
\value{
The function returns the evaluation of the target function,
which is used to decide whether to accept or reject proposed states
in the MH algorithm. Returns NA if there are no gamma parameters.
}
\description{
\code{initial_omega_j} computes OLS quantity for initial omega_j using initialized
gamma_j'
}
\keyword{internal}
