% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koma-package.R
\docType{package}
\name{koma}
\alias{koma}
\alias{koma-package}
\title{koma: Large Scale Macroeconomic Model}
\description{
Tools for estimating and forecasting large-scale macroeconomic models,
including extended time series utilities used throughout the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://timothymerlin.github.io/koma/}
  \item \url{https://github.com/TimothyMerlin/koma}
  \item Report bugs at \url{https://github.com/TimothyMerlin/koma/issues}
}

}
\author{
\strong{Maintainer}: Merlin Scherer \email{scherer@kof.ethz.ch} (\href{https://orcid.org/0009-0005-7479-168X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Samad Sarferaz \email{sarferaz@kof.ethz.ch} [copyright holder]
  \item Laurent Florin \email{florin@kof.ethz.ch}
}

Other contributors:
\itemize{
  \item Andrew D. Martin (Adapted rwish/riwish code from MCMCpack) [contributor, copyright holder]
  \item Kevin M. Quinn (Adapted rwish/riwish code from MCMCpack) [contributor, copyright holder]
  \item Jong Hee Park (Adapted rwish/riwish code from MCMCpack) [contributor, copyright holder]
  \item Bill Venables (Adapted multivariate_norm code from MASS::mvrnorm) [contributor, copyright holder]
  \item Brian D. Ripley (Adapted multivariate_norm code from MASS::mvrnorm) [contributor, copyright holder]
}

}
