% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koma-package.R
\docType{package}
\name{koma}
\alias{koma}
\alias{koma-package}
\title{koma: Large Scale Macroeconomic Model}
\description{
Tools for estimating and forecasting large-scale macroeconomic models,
including extended time series utilities used throughout the package.
}
\author{
\strong{Maintainer}: Merlin Scherer \email{scherer@kof.ethz.ch} (\href{https://orcid.org/0009-0005-7479-168X}{ORCID})

Authors:
\itemize{
  \item Samad Sarferaz \email{sarferaz@kof.ethz.ch} [copyright holder]
  \item Laurent Florin \email{florin@kof.ethz.ch}
}

}
