% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_of_equations.R
\name{print.koma_seq}
\alias{print.koma_seq}
\title{Print System of Equations}
\usage{
\method{print}{koma_seq}(x, ...)
}
\arguments{
\item{x}{An object of class \code{koma_seq}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
This function prints an object of class \code{koma_seq}
to ensure the equations are displayed in a readable format.
}
