% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{summary.koma_estimate}
\alias{summary.koma_estimate}
\title{Summary method for koma_estimate objects}
\usage{
\method{summary}{koma_estimate}(object, ...)
}
\arguments{
\item{object}{A koma_estimate object.}

\item{...}{Additional parameters:
\describe{
\item{variables}{Optional. A character vector of variables to summarize.
Default is NULL, which means all variables will be summarized.}
\item{central_tendency}{Optional. A string specifying the measure of
central tendency ("mean", "median"). Default is "mean".}
\item{ci_low}{Optional. Lower bound of the confidence interval. Default
is 5.}
\item{ci_up}{Optional. Upper bound of the confidence interval. Default is
95.}
\item{texreg_object}{Optional. If TRUE, returns a texreg object. Default
is FALSE, which returns an ASCII table.}
\item{digits}{Optional. Number of digits to round numeric values.
Default is 2.}
}}
}
\value{
Depending on the value of texreg_object, returns either a list of
texreg objects or prints an ASCII table.
}
\description{
This function provides a summary for koma_estimate objects.
It can return either a texreg object or an ASCII table.
}
