% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_posterior.R
\name{update_estimates_with_weights}
\alias{update_estimates_with_weights}
\title{Update Gamma and Beta Matrices with Identity Weights}
\usage{
update_estimates_with_weights(identities, gamma_matrix, beta_matrix)
}
\arguments{
\item{identities}{A list of identity equations.}

\item{gamma_matrix}{Initial gamma matrix derived from median estimates.}

\item{beta_matrix}{Initial beta matrix derived from median estimates.}
}
\value{
A list containing:
\itemize{
\item \code{gamma_matrix}: Gamma matrix updated with identity weights.
\item \code{beta_matrix}: Beta matrix updated with identity weights.
}
}
\description{
The function takes in the identities and initial matrices for gamma and beta,
and updates the values in the matrices based on the identity weights.
}
\seealso{
\code{\link{construct_posterior}} for a function that uses this function's
outputs.
}
\keyword{internal}
